bot/requirements.txt
aiogram
torch
torchvision
pillow
numpy
